Almost complete splitting BitcoindRpcClient into multitple traits
