+++ rpc/src/main/scala/org/bitcoins/rpc/client/MessageRpc.scala
@@ -7,9 +7,9 @@ import play.api.libs.json.JsString
 import scala.concurrent.Future
 
 /**
-  * RPC calls related to the message signing functionality
-  * in bitcoind
-  */
+ * RPC calls related to the message signing functionality
+ * in bitcoind
+ */
 protected trait MessageRpc extends Client {
   def signMessage(address: P2PKHAddress, message: String): Future[String] = {
     bitcoindCall[String](
@@ -18,17 +18,17 @@ protected trait MessageRpc extends Client {
   }
 
   def signMessageWithPrivKey(
-                              key: ECPrivateKey,
-                              message: String): Future[String] = {
+    key: ECPrivateKey,
+    message: String): Future[String] = {
     bitcoindCall[String](
       "signmessagewithprivkey",
       List(JsString(key.toWIF(network)), JsString(message)))
   }
 
   def verifyMessage(
-                     address: P2PKHAddress,
-                     signature: String,
-                     message: String): Future[Boolean] = {
+    address: P2PKHAddress,
+    signature: String,
+    message: String): Future[Boolean] = {
     bitcoindCall[Boolean](
       "verifymessage",
       List(JsString(address.value), JsString(signature), JsString(message)))
