+++ rpc/src/main/scala/org/bitcoins/rpc/client/MultisigRpc.scala
@@ -4,18 +4,18 @@ import org.bitcoins.core.crypto.ECPublicKey
 import org.bitcoins.core.protocol.P2PKHAddress
 import org.bitcoins.rpc.client.RpcOpts.AddressType
 import org.bitcoins.rpc.jsonmodels.MultiSigResult
-import org.bitcoins.rpc.serializers.JsonSerializers._
-import play.api.libs.json.{JsArray, JsNumber, JsString, Json}
+import org.bitcoins.rpc.serializers.BitcoindJsonSerializers._
+import play.api.libs.json.{ JsArray, JsNumber, JsString, Json }
 
 import scala.concurrent.Future
 
 protected trait MultisigRpc extends Client {
 
   private def addMultiSigAddress(
-                                  minSignatures: Int,
-                                  keys: Vector[Either[ECPublicKey, P2PKHAddress]],
-                                  account: String = "",
-                                  addressType: Option[AddressType]): Future[MultiSigResult] = {
+    minSignatures: Int,
+    keys: Vector[Either[ECPublicKey, P2PKHAddress]],
+    account: String = "",
+    addressType: Option[AddressType]): Future[MultiSigResult] = {
     def keyToString(key: Either[ECPublicKey, P2PKHAddress]): JsString =
       key match {
         case Right(k) => JsString(k.value)
@@ -40,32 +40,32 @@ protected trait MultisigRpc extends Client {
   }
 
   def addMultiSigAddress(
-                          minSignatures: Int,
-                          keys: Vector[Either[ECPublicKey, P2PKHAddress]]): Future[MultiSigResult] =
+    minSignatures: Int,
+    keys: Vector[Either[ECPublicKey, P2PKHAddress]]): Future[MultiSigResult] =
     addMultiSigAddress(minSignatures, keys, addressType = None)
 
   def addMultiSigAddress(
-                          minSignatures: Int,
-                          keys: Vector[Either[ECPublicKey, P2PKHAddress]],
-                          account: String): Future[MultiSigResult] =
+    minSignatures: Int,
+    keys: Vector[Either[ECPublicKey, P2PKHAddress]],
+    account: String): Future[MultiSigResult] =
     addMultiSigAddress(minSignatures, keys, account, None)
 
   def addMultiSigAddress(
-                          minSignatures: Int,
-                          keys: Vector[Either[ECPublicKey, P2PKHAddress]],
-                          addressType: AddressType): Future[MultiSigResult] =
+    minSignatures: Int,
+    keys: Vector[Either[ECPublicKey, P2PKHAddress]],
+    addressType: AddressType): Future[MultiSigResult] =
     addMultiSigAddress(minSignatures, keys, addressType = Some(addressType))
 
   def addMultiSigAddress(
-                          minSignatures: Int,
-                          keys: Vector[Either[ECPublicKey, P2PKHAddress]],
-                          account: String,
-                          addressType: AddressType): Future[MultiSigResult] =
+    minSignatures: Int,
+    keys: Vector[Either[ECPublicKey, P2PKHAddress]],
+    account: String,
+    addressType: AddressType): Future[MultiSigResult] =
     addMultiSigAddress(minSignatures, keys, account, Some(addressType))
 
   def createMultiSig(
-                      minSignatures: Int,
-                      keys: Vector[ECPublicKey]): Future[MultiSigResult] = {
+    minSignatures: Int,
+    keys: Vector[ECPublicKey]): Future[MultiSigResult] = {
     bitcoindCall[MultiSigResult](
       "createmultisig",
       List(JsNumber(minSignatures), Json.toJson(keys.map(_.hex))))
