+++ rpc/src/main/scala/org/bitcoins/rpc/serializers/BitcoindJsonWriters.scala
@@ -7,9 +7,13 @@ import org.bitcoins.core.protocol.BitcoinAddress
 import org.bitcoins.core.protocol.script.ScriptPubKey
 import org.bitcoins.core.protocol.transaction.{ Transaction, TransactionInput }
 import org.bitcoins.core.util.BitcoinSUtil
+import org.bitcoins.rpc.client.RpcOpts
 import play.api.libs.json._
 
-object JsonWriters {
+import scala.language.implicitConversions
+
+object BitcoindJsonWriters {
+
   implicit object BitcoinsWrites extends Writes[Bitcoins] {
     override def writes(o: Bitcoins): JsValue = JsNumber(o.toBigDecimal)
   }
@@ -51,4 +55,10 @@ object JsonWriters {
       Json.toJson(o.map { case (k, v) => (keyString(k), v) })
     }
   }
+
+  implicit val importMultiAddressWrites: Writes[RpcOpts.ImportMultiAddress] =
+    Json.writes[RpcOpts.ImportMultiAddress]
+  implicit val importMultiRequestWrites: Writes[RpcOpts.ImportMultiRequest] =
+    Json.writes[RpcOpts.ImportMultiRequest]
 }
+
